refactor: context files atoms to context
