+++ components/01-atoms/NetworkDropdown.tsx
@@ -4,7 +4,6 @@ import {
   ArrowIconVariant,
   NetworkIcon,
   NetworkVariants,
-  SwapContext,
   Tooltip,
 } from "@/components/01-atoms";
 import {
@@ -15,6 +14,7 @@ import {
 import { useAuthenticatedUser } from "@/lib/client/hooks/useAuthenticatedUser";
 import { useSupportedNetworks } from "@/lib/client/hooks/useSupportedNetworks";
 import { capitalizeFirstLetterPrhases } from "@/lib/client/utils";
+import { SwapContext } from "@/lib/client/contexts";
 import { useSwitchNetwork, sepolia, useNetwork } from "wagmi";
 import React, { useContext, useEffect, useState } from "react";
 import toast from "react-hot-toast";
