+++ components/01-atoms/index.ts
@@ -14,7 +14,6 @@ export * from "./ProgressBar";
 export * from "./SearchBar";
 export * from "./SearchItemsShelf";
 export * from "./StatusOffers";
-export * from "./SwapContext";
 export * from "./SwapExpireTime";
 export * from "./SwapModalButton";
 export * from "./SwapModalLayout";
@@ -26,7 +25,6 @@ export * from "./TokenOfferDetails";
 export * from "./TokensOfferSkeleton";
 export * from "./TokensShelfTab";
 export * from "./Tooltip";
-export * from "./OffersContext";
 
 /** Icons */
 export * from "./icons/ArrowIcon";
