+++ components/02-molecules/OfferSummary.tsx
@@ -1,11 +1,7 @@
 import { ForWhom } from "../03-organisms";
-import {
-  ENSAvatar,
-  ENSAvatarSize,
-  PersonIcon,
-  SwapContext,
-} from "@/components/01-atoms";
+import { ENSAvatar, ENSAvatarSize, PersonIcon } from "@/components/01-atoms";
 import { TokenCardStyleType, TokensList } from "@/components/02-molecules";
+import { SwapContext } from "@/lib/client/contexts";
 import { useAuthenticatedUser } from "@/lib/client/hooks/useAuthenticatedUser";
 import { useEnsData } from "@/lib/client/hooks/useENSData";
 import { useContext } from "react";
