+++ components/02-molecules/TheSidebarHeader.tsx
@@ -1,5 +1,5 @@
 import { CloseCTA, DisconnectWallet } from "@/components/01-atoms";
-import { useSidebar } from "@/lib/client/contexts/SidebarContext.tsx";
+import { useSidebar } from "@/lib/client/contexts/SidebarContext";
 import React from "react";
 import cc from "classcat";
 import { useTheme } from "next-themes";
