+++ components/03-organisms/SwapStation.tsx
@@ -1,6 +1,5 @@
 import {
   PaperPlane,
-  SwapContext,
   SwapExpireTime,
   SwapIcon,
   SwapIconVariant,
@@ -13,6 +12,7 @@ import {
 } from "@/components/02-molecules";
 import { ForWhom } from "@/components/03-organisms";
 import { useAuthenticatedUser } from "@/lib/client/hooks/useAuthenticatedUser";
+import { SwapContext } from "@/lib/client/contexts";
 import { useContext, useEffect, useState } from "react";
 import cc from "classcat";
 import { toast } from "react-hot-toast";
