+++ components/03-organisms/TokensShelf.tsx
@@ -5,10 +5,11 @@ import {
   getERC20TokensFromAddress,
 } from "@/lib/client/blockchain-utils";
 import { TokensList } from "@/components/02-molecules";
-import { SelectUserIcon, SwapContext } from "@/components/01-atoms";
+import { SelectUserIcon } from "@/components/01-atoms";
 import { useSupportedNetworks } from "@/lib/client/hooks/useSupportedNetworks";
 import { Token } from "@/lib/shared/types";
 import { ShelfContext } from "@/lib/client/contexts/ShelfContext";
+import { SwapContext } from "@/lib/client/contexts";
 import { useContext, useEffect, useState } from "react";
 import { useNetwork } from "wagmi";
 /* eslint-disable react-hooks/exhaustive-deps */
