+++ lib/client/contexts/SidebarContext.tsx
